Fixing issue with queue cards not appearing due to earlier refactor
